backend/requirements.txt
gradio>=3.50.0,<4.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pillow>=10.0.0
numpy>=1.24.0
python-multipart>=0.0.6
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
safetensors
diff-gaussian-rasterization
simple-knn